requirements.txt
elasticsearch==8.15.1
fastapi==0.115.0
langchain==0.3.2
langchain_core==0.3.9
langchain_google_genai==2.0.0
langchain_openai==0.2.2
protobuf==5.28.2
pydub==0.25.1
PyPDF2==3.0.1
python-dotenv==1.0.1
Requests==2.32.3
streamlit==1.39.0
uvicorn==0.31.0